fix(contracts): widen tier-threshold type from uint8 to uint64
Discovered while writing Escrow.t.sol that the spec's own default
tier curve [0, 1, 3, 10, 30, 100, 300, 1000] does not fit in uint8
(max 255). Promoting to uint64 also matches the return type of
ReputationRegistry.getSummary().count, eliminating implicit casts in
releasedBps() and walking the tier table directly without zero-extension.

Cost: marginal — the array is stored once at deploy and re-walked on
every withdraw, but width never changes computation depth.

Updated:
- specs/09-l4d-escrow.md (constructor sig + tierConfig sig + factory sig)
- contracts/src/Escrow.sol (state var, ctor, event, tierConfig)
- contracts/src/EscrowFactory.sol (event, createEscrow, predictAddress)

+++ contracts/src/Escrow.sol
@@ -36,7 +36,9 @@ contract Escrow is ReentrancyGuard {
 
     // Tier curve, pinned per Escrow at deploy. Stored as plain dynamic
     // arrays + two strings; constructor validates monotonicity + bounds.
-    uint8[]  private _tierThresholds;
+    // Threshold type is uint64 to match ReputationRegistry.getSummary's
+    // return type (avoids implicit casts during ramp evaluation).
+    uint64[] private _tierThresholds;
     uint16[] private _tierReleaseBps;
     string   private _tag1;
     string   private _tag2;
@@ -54,7 +56,7 @@ contract Escrow is ReentrancyGuard {
         address indexed reputationRegistry,
         uint256          agentId,
         address          facilitatorClient,
-        uint8[]          tierThresholds,
+        uint64[]         tierThresholds,
         uint16[]         tierReleaseBps,
         string           tag1,
         string           tag2
@@ -75,7 +77,7 @@ contract Escrow is ReentrancyGuard {
         address          reputationRegistry_,
         uint256          agentId_,
         address          facilitatorClient_,
-        uint8[]   memory tierThresholds_,
+        uint64[]  memory tierThresholds_,
         uint16[]  memory tierReleaseBps_,
         string    memory tag1_,
         string    memory tag2_
@@ -251,7 +253,7 @@ contract Escrow is ReentrancyGuard {
         external
         view
         returns (
-            uint8[]  memory thresholds,
+            uint64[] memory thresholds,
             uint16[] memory releaseBpsArr,
             string   memory tag1Val,
             string   memory tag2Val
